requirements.txt
streamlit
openai-whisper
yt-dlp
torch
ffmpeg-python
st_copy_to_clipboard